# server/src/main/java/com/github/KinConnect/controllers/AuthController.java
package com.github.KinConnect.controllers;

import com.github.KinConnect.dto.Response;
import com.github.KinConnect.dto.auth.UserLoginDto;
import com.github.KinConnect.dto.auth.UserLoginResponse;
import com.github.KinConnect.dto.auth.UserRegisterDto;
import com.github.KinConnect.dto.auth.UserVerifyDto;
import com.github.KinConnect.services.UserService;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

/**
 * @author yihangz
 */
@RestController
@RequestMapping("/auth")
@RequiredArgsConstructor
public class AuthController {

    private final UserService userService;


    @PostMapping("/register")
    public ResponseEntity<Response> register(@Valid @RequestBody UserRegisterDto userRegisterDto) {
        // register the user and send the verification code
        userService.newUser(userRegisterDto.getEmail(), userRegisterDto.getUsername(), userRegisterDto.getPassword());
        return ResponseEntity.status(HttpStatus.CREATED).body(Response.builder().code(201).message("Registration successful").build());

    }

    @PostMapping("/verify-email")
    public ResponseEntity<Response> verifyEmail(@Valid @RequestBody UserVerifyDto userVerifyDto) {
        userService.verifyUser(userVerifyDto.getEmail(), userVerifyDto.getCode());
        return ResponseEntity.status(HttpStatus.CREATED).body(Response.builder().code(201).message("Verification successful").build());

    }

    @PostMapping("/login")
    public ResponseEntity<Response> login(@Valid @RequestBody UserLoginDto userLoginDto) {
        String jwtToken = userService.login(userLoginDto.getEmail(), userLoginDto.getPassword());
        return ResponseEntity.status(HttpStatus.OK).body(Response.builder().code(200).message("Login successful").data(new UserLoginResponse(jwtToken)).build());

    }
}