# server/src/main/java/com/github/KinConnect/doc/api.openapi.yaml
openapi: 3.0.3
info:
  title: KinConnect
  description: For kin to connect
  version: 1.0.2
servers:
  - url: https://api.kinconnect.com

paths:
  /auth/register:
    post:
      summary: 用户注册
      description: 注册一个新用户账号（需进行邮箱验证）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - username
              properties:
                email:
                  type: string
                  format: email
                  example: han@example.com
                password:
                  type: string
                  format: password
                  example: myStrongPassword123
                username:
                  type: string
                  example: oosomnus
      responses:
        '201':
          description: 注册成功，验证邮件已发送
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    example: 64f21a9b3e
                  message:
                    type: string
                    example: 注册成功，请查收验证邮件
        '400':
          description: 请求参数错误
        '409':
          description: 邮箱已被注册

  /auth/verify-email:
    post:
      summary: 邮箱验证
      description: 用户通过邮件中的验证码完成邮箱验证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - code
              properties:
                email:
                  type: string
                  format: email
                  example: han@example.com
                code:
                  type: string
                  description: 邮箱验证码
                  example: 123456
      responses:
        '200':
          description: 邮箱验证成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 邮箱验证成功
        '400':
          description: 验证码错误或已过期
        '404':
          description: 邮箱未找到

  /auth/login:
    post:
      summary: 用户登录（邮箱）
      description: 使用邮箱和密码登录，需先完成邮箱验证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: han@example.com
                password:
                  type: string
                  format: password
                  example: myStrongPassword123
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT 令牌
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: 邮箱或密码错误
        '403':
          description: 邮箱未验证
        '404':
          description: 邮箱未注册
  /user/switch-role:
    post:
      summary: 用户切换角色
      description: 切换现有角色 (0/A)
      security:
        - bearerAuth: [ ]
      responses:
        '201':
          description: 切换成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: "切换成功"
                  data:
                    type: object
        '400':
          description: 切换失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "切换失败"
  /messages/update/{old_id}:
    post:
      summary: 增加/更改/删除现有消息
      description: 根据 `client_id` 批量更新消息内容，可包括新增、修改或删除操作。
      parameters:
        - name: old_id
          in: path
          required: true
          schema:
            type: string
          description: 客户端唯一标识
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                msg:
                  type: array
                  description: 消息对象数组
                  items:
                    $ref: '#/components/schemas/Message'
      responses:
        '201':
          description: 消息更新成功

  /messages/{old_id}:
    get:
      summary: 获取指定客户端的消息列表
      description: 根据 `old_id` 查询当前关联的所有消息。
      parameters:
        - name: old_id
          in: path
          required: true
          schema:
            type: string
          description: 客户端唯一标识
      responses:
        '200':
          description: 返回消息数组
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'

  /user/info:
    get:
      summary: 获取用户信息
      description: 获取用户信息
      responses:
        '200':
          description: 返回用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 消息唯一标识
          example: "550e8400-e29b-41d4-a716-446655440000"
        execTime:
          type: string
          description: Cron 表达式，定义消息执行时间
          example: "0 0 9 * * ?"
        text:
          type: string
          description: 消息文本内容
          maxLength: 500
          example: "每日九点发送的提醒消息"
        isOneTime:
          type: boolean
          description: 是否仅执行一次
          example: true

    User:
      type: object
      description: 用户实体对象
      properties:
        id:
          type: integer
          format: int64
          description: 用户唯一标识
          example: 1
        username:
          type: string
          description: 用户名
          example: "Han"
        email:
          type: string
          format: email
          description: 用户邮箱地址
          example: "han@example.com"
        isOld:
          type: boolean
          description: 是否为被照护老人账户
          example: false
        guardian:
          $ref: '#/components/schemas/UserSimple'
        olds:
          type: array
          description: 被当前监护用户（guardian）关联的老人列表
          items:
            $ref: '#/components/schemas/UserSimple'
        isVerified:
          type: boolean
          description: 邮箱是否已验证
          example: true
      required:
        - id
        - username
        - email
        - isVerified

    UserSimple:
      type: object
      description: 简化用户对象（用于关联字段）
      properties:
        id:
          type: integer
          format: int64
          example: 2
        username:
          type: string
          example: "ElderLee"
        email:
          type: string
          format: email
          example: "elderlee@example.com"